server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
litellm==1.52.15
aiohttp==3.11.8
pydantic==2.10.2